requirements.txt
# Core + drdroid-debug-toolkit (install with: pip install -r requirements.txt)
mcp
python-dotenv
git+https://github.com/DrDroidLab/drdroid-debug-toolkit.git@master
